refactor(*): merge code to marginally speed up compilation
It's not exceedingly helpful to have all of these types in their
own files anyway, and it improves compile time.
