+++ src/win32/kerberos_win32.cc
@@ -1,9 +1,6 @@
 #include <memory>
 
 #include "../kerberos.h"
-#include "../kerberos_client.h"
-#include "../kerberos_common.h"
-#include "../kerberos_server.h"
 #include "../kerberos_worker.h"
 
 #define GSS_MECH_OID_KRB5 9
@@ -31,6 +28,108 @@ NAN_INLINE std::wstring WStringOptionValue(v8::Local<v8::Object> options, const 
                : std::wstring();
 }
 
+/// KerberosClient
+KerberosClient::~KerberosClient() {
+    // if (_state != NULL) {
+    //     authenticate_gss_client_clean(_state);
+    //     _state = NULL;
+    // }
+}
+
+NAN_METHOD(KerberosClient::Step) {
+    KerberosClient* client = Nan::ObjectWrap::Unwrap<KerberosClient>(info.This());
+    std::string challenge(*Nan::Utf8String(info[0]));
+    Nan::Callback* callback = new Nan::Callback(Nan::To<v8::Function>(info[1]).ToLocalChecked());
+
+    KerberosWorker::Run(callback, "kerberos:ClientStep", [=](KerberosWorker::SetOnFinishedHandler onFinished) {
+        std::shared_ptr<sspi_result> result(
+            auth_sspi_client_step(client->state(), (SEC_CHAR*)challenge.c_str(), NULL), ResultDeleter);
+
+        return onFinished([=](KerberosWorker* worker) {
+            Nan::HandleScope scope;
+            if (result->code == AUTH_GSS_ERROR) {
+                v8::Local<v8::Value> argv[] = {Nan::New(result->message).ToLocalChecked(), Nan::Null()};
+                worker->Call(2, argv);
+                return;
+            }
+
+            v8::Local<v8::Value> response = Nan::Null();
+            if (client->state()->response != NULL) {
+                response = Nan::New(client->state()->response).ToLocalChecked();
+            }
+
+            v8::Local<v8::Value> argv[] = {Nan::Null(), response};
+            worker->Call(2, argv);
+        });
+    });
+}
+
+NAN_METHOD(KerberosClient::UnwrapData) {
+    KerberosClient* client = Nan::ObjectWrap::Unwrap<KerberosClient>(info.This());
+    std::string challenge(*Nan::Utf8String(info[0]));
+    Nan::Callback* callback = new Nan::Callback(Nan::To<v8::Function>(info[1]).ToLocalChecked());
+
+    KerberosWorker::Run(callback, "kerberos:ClientUnwrap", [=](KerberosWorker::SetOnFinishedHandler onFinished) {
+        std::shared_ptr<sspi_result> result(
+            auth_sspi_client_unwrap(client->state(), (SEC_CHAR*)challenge.c_str()), ResultDeleter);
+
+        return onFinished([=](KerberosWorker* worker) {
+            Nan::HandleScope scope;
+            if (result->code == AUTH_GSS_ERROR) {
+                v8::Local<v8::Value> argv[] = {Nan::New(result->message).ToLocalChecked(), Nan::Null()};
+                worker->Call(2, argv);
+                return;
+            }
+
+            v8::Local<v8::Value> argv[] = {Nan::Null(), Nan::New(client->state()->response).ToLocalChecked()};
+            worker->Call(2, argv);
+        });
+    });
+}
+
+NAN_METHOD(KerberosClient::WrapData) {
+    KerberosClient* client = Nan::ObjectWrap::Unwrap<KerberosClient>(info.This());
+    std::string challenge(*Nan::Utf8String(info[0]));
+    v8::Local<v8::Object> options = Nan::To<v8::Object>(info[1]).ToLocalChecked();
+    Nan::Callback* callback = new Nan::Callback(Nan::To<v8::Function>(info[2]).ToLocalChecked());
+    std::string user = StringOptionValue(options, "user");
+    int protect = 0; // NOTE: this should be an option
+
+    KerberosWorker::Run(callback, "kerberos:ClientWrap", [=](KerberosWorker::SetOnFinishedHandler onFinished) {
+        std::shared_ptr<sspi_result> result(auth_sspi_client_wrap(
+            client->state(), (SEC_CHAR*)challenge.c_str(), (SEC_CHAR*)user.c_str(), user.length(), protect), ResultDeleter);
+
+        return onFinished([=](KerberosWorker* worker) {
+            Nan::HandleScope scope;
+
+            if (result->code == AUTH_GSS_ERROR) {
+                v8::Local<v8::Value> argv[] = {Nan::New(result->message).ToLocalChecked(), Nan::Null()};
+                worker->Call(2, argv);
+                return;
+            }
+
+            v8::Local<v8::Value> argv[] = {Nan::Null(), Nan::New(client->state()->response).ToLocalChecked()};
+            worker->Call(2, argv);
+        });
+    });
+}
+
+/// KerberosServer
+KerberosServer::~KerberosServer() {
+    // if (_state != NULL) {
+    //     authenticate_gss_server_clean(_state);
+    //     _state = NULL;
+    // }
+}
+
+NAN_METHOD(KerberosServer::Step) {
+    KerberosServer* server = Nan::ObjectWrap::Unwrap<KerberosServer>(info.This());
+    std::string challenge(*Nan::Utf8String(info[0]));
+    Nan::Callback* callback = new Nan::Callback(Nan::To<v8::Function>(info[1]).ToLocalChecked());
+    Nan::ThrowError("`KerberosServer::Step` is not implemented yet for windows");
+}
+
+/// Global Methods
 NAN_METHOD(InitializeClient) {
     std::wstring service(to_wstring(v8::String::Utf8Value(info[0]->ToString())));
     v8::Local<v8::Object> options = Nan::To<v8::Object>(info[1]).ToLocalChecked();
